+++ xlMemoryPool.h
@@ -27,10 +27,10 @@ public:
 	MemoryPool(void);
 	~MemoryPool(void);
 	void ClearPool(void);
-	LPSTR GetTempMemory(int cBytes);
+	LPSTR GetTempMemory(size_t cBytes);
 	void FreeAllTempMemory();
 
 	DWORD m_dwOwner;			// ID of ownning thread
 	char* m_rgchMemBlock;		// Memory for temporary XLOPERs
-	int m_ichOffsetMemBlock;	// Offset of next memory block to allocate
+	size_t m_ichOffsetMemBlock;	// Offset of next memory block to allocate
 };
